move completer into widgets
